Correction

+++ Octal->Decimal.cpp
@@ -5,7 +5,7 @@
 #include <cmath>
 using namespace std;
 
-int Convert(long long num)
+int Convert(int num)
 {
     int r, total = 0, i = 0;
     while (num != 0)
@@ -21,7 +21,7 @@ int Convert(long long num)
 
 int main()
 {
-    long long n;
+    int n;
     cin >> n;
     cout << n << " in octal = " << Convert(n) << " in decimal";
     return 0;
